refactor: move lab changed-since preflight
