+++ src/core/runner/mod.rs
@@ -13,6 +13,7 @@ mod apply;
 mod connection;
 mod evidence;
 mod execution;
+mod offload_changed_since;
 mod offload_metadata;
 mod workspace;
 
@@ -29,6 +30,7 @@ pub use evidence::{
 };
 pub(crate) use execution::daemon_api_get;
 pub use execution::{exec, RunnerExecMode, RunnerExecOptions, RunnerExecOutput};
+pub use offload_changed_since::preflight_lab_offload_changed_since;
 pub use offload_metadata::{capture_lab_offload_metadata, lab_offload_metadata};
 pub use workspace::{
     sync_workspace, RunnerWorkspaceSyncMode, RunnerWorkspaceSyncOptions, RunnerWorkspaceSyncOutput,
